feat: Implement parent-child sync for all accessible customers (#45)
* feat(): fix the parent-child logic

At the moment `AccessibleCustomers` is not being used as a parent stream to `CustomerHierarchyStream`. This update fixes that. By updating to the latest singer-sdk version we can use `generate_child_contexts` to generate multiple child streams per parent stream to sync all accessible customers.

I also added a check to make sure a customer is enabled to avoid 403 errors.

Also changed some of the tap config because in my opinion it was a bit unclear. `login_customer_id` is now required as it is required in the header. You can also specify `customer_ids` to filter on customers you want to sync. But only if you have access to them and they are enabled. I understand that some people need to update their config but in the end I think this is more clear.

Moved pagination to the new singer-sdk pagination setup: https://sdk.meltano.com/en/v0.40.0/guides/pagination-classes.html

Other updates are linting updates.

* Update tap_googleads/tap.py

Co-authored-by: ReubenFrankel <60552974+ReubenFrankel@users.noreply.github.com>

* Update tap_googleads/streams.py

Co-authored-by: ReubenFrankel <60552974+ReubenFrankel@users.noreply.github.com>

* Maintain support for `customer_id` config

* Remove unused import

* Update README with info on `customer_ids`/`customer_id`

* Fix tests broken by SDK version bump and config

* Add test workflow

* Update locked dependencies

---------

Co-authored-by: ReubenFrankel <60552974+ReubenFrankel@users.noreply.github.com>
Co-authored-by: Reuben Frankel <rfrankel@matatika.com>